Stock Adjustment Done. Sp change send as a Whatsapp

+++ PointOfSale.Core/Interfaces/Repositories/Inventory/IStockAdjustmentRepository.cs
@@ -7,7 +7,7 @@ namespace PointOfSale.Core.Interfaces.Repositories.Inventory
 {
     public interface IStockAdjustmentRepository
     {
-        Task<int> CreateAsync(StockAdjustment stockAdjustment);
+        Task<long> CreateAsync(StockAdjustment stockAdjustment);
         Task<IEnumerable<StockAdjustment>> GetAllAsync();
         Task<IEnumerable<AvailableQuantityDto>> GetAvailableQty(int locationId, int productId);
     }
